+++ Entity Component System/src/at/fhooe/mtd/ecs/Engine.java
@@ -250,7 +250,9 @@ public final class Engine {
         
         // update systems
         for (EngineSystem s : systems) {
-            s.update(dt);
+        	if (s.isEnabled()) {
+                s.update(dt);        		
+        	}
         }
         
         // execute pending commands